# Conflicts:
#	src/main/java/school/faang/user_service/service/event/EventService.java

+++ src/main/java/school/faang/user_service/service/event/EventService.java
@@ -50,26 +50,22 @@ public class EventService {
 
     }
 
-    public EventDto getEvent(long id){
-        Optional<Event> event;
+    public EventDto getEvent(long id) {
 
-        try {
-            event = eventRepository.findById(id);
-        } catch (IllegalArgumentException exception) {
+        if (id <= 0) {
             throw new DataValidationException("ID is incorrect");
         }
-        if (event.isEmpty()) {
-            throw new DataValidationException("There is no event with this id");
-        }
 
-        return eventMapper.toDTO(event.get());
+        return eventMapper.toDTO(
+                eventRepository
+                        .findById(id)
+                        .orElseThrow(() -> new DataValidationException("There is no event with this id"))
+        );
     }
 
     public void deleteEvent(long id) {
-        try {
+        if (id > 0) {
             eventRepository.deleteById(id);
-        } catch (IllegalArgumentException exception) {
-            throw new DataValidationException("ID is incorrect");
         }
     }
